Add UserInputsException and validation

+++ src/main/java/com/amalitech/SpringBootBloggingApp/service/impl/ReviewServiceImpl.java
@@ -6,6 +6,8 @@ import com.amalitech.SpringBootBloggingApp.model.entity.User;
 import com.amalitech.SpringBootBloggingApp.repository.impl.ReviewRepositoryImpl;
 import com.amalitech.SpringBootBloggingApp.service.ReviewService;
 
+import com.amalitech.SpringBootBloggingApp.util.ValidationUtil;
+import com.amalitech.SpringBootBloggingApp.util.exceptions.UserInputsException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,26 +24,50 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public Review create(Review review) {
+        if (!ValidationUtil.isValidObjectId(review.getUserId())) {
+            throw new UserInputsException("Invalid user ID");
+        }
+        if (!ValidationUtil.isValidObjectId(review.getPostId())) {
+            throw new UserInputsException("Invalid post ID");
+        }
+        if (!ValidationUtil.isValidRating(review.getRating())) {
+            throw new UserInputsException("Invalid rating");
+        }
+        if (!ValidationUtil.isValidContent(review.getFeedback())) {
+            throw new UserInputsException("Invalid content");
+        }
         return reviewRepository.save(review);
     }
 
     @Override
     public boolean delete(String reviewId) {
+        if (!ValidationUtil.isValidObjectId(reviewId)) {
+            throw new UserInputsException("Invalid review ID");
+        }
         return reviewRepository.deleteById(reviewId);
     }
 
     @Override
     public boolean update(Review review) {
+        if (!ValidationUtil.isValidObjectId(review.getId())) {
+            throw new UserInputsException("Invalid review ID");
+        }
         return reviewRepository.update(review);
     }
 
     @Override
     public List<Review> getByPost(String postId) {
+        if (!ValidationUtil.isValidObjectId(postId)) {
+            throw new UserInputsException("Invalid post ID");
+        }
         return reviewRepository.findByPostId(postId);
     }
 
     @Override
     public double getAverageRatingForPost(String postId) {
+        if (!ValidationUtil.isValidObjectId(postId)) {
+            throw new UserInputsException("Invalid post ID");
+        }
         return reviewRepository.calculateAverageRating(postId);
     }
